+++ osu.Game/Screens/Multi/Components/MatchBeatmapDetailArea.cs
@@ -54,6 +54,7 @@ namespace osu.Game.Screens.Multi.Components
                             new TriangleButton
                             {
                                 Text = "添加该谱面至游玩列表",
+                                FontSize = 22,
                                 RelativeSizeAxes = Axes.Both,
                                 Size = Vector2.One,
                                 Action = () => CreateNewItem?.Invoke()